+++ sketch_map_tool/config.py
@@ -23,10 +23,10 @@ def load_config_default() -> Dict[str, str]:
         "data-dir": get_default_data_dir(),
         "user-agent": "sketch-map-tool",
         "broker-url": "redis://localhost:6379",
-        "result-backend": "redis://localhost:6379",
+        "result-backend": "db+postgresql://smt:smt@localhost:5432",
         "data-store": "redis://localhost:6379",
-        "wms-url": "https://maps.heigit.org/osm-wms/service?SERVICE=WMS&VERSION=1.1.1",
-        "wms-layers": "osm_auto:all",
+        "wms-url": "https://maps.heigit.org/osm-carto/service?SERVICE=WMS&VERSION=1.1.1",
+        "wms-layers": "heigit:osm-carto@2xx",
     }