+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 - add content and style to index ([#73])
 - use celery as queue to run background tasks ([#72])
 - implement web map service client ([#95])
+- add module containing the paper format specs ([#99])
+- include map scale as request parameter to the backend ([#107])
+- add module to generate qr-codes as svg & rlg ([#98])
+- add resources (compass and globes as SVG graphics) for map generation ([#101])
+- use postgres as celery result backend ([#116)]
 
 ### Bug Fixes
 
@@ -37,4 +42,9 @@
 [#72]: https://github.com/GIScience/sketch-map-tool/pull/72
 [#73]: https://github.com/GIScience/sketch-map-tool/pull/73
 [#95]: https://github.com/GIScience/sketch-map-tool/pull/95
+[#98]: https://github.com/GIScience/sketch-map-tool/pull/98
+[#99]: https://github.com/GIScience/sketch-map-tool/pull/99
+[#101]: https://github.com/GIScience/sketch-map-tool/pull/101
 [#102]: https://github.com/GIScience/sketch-map-tool/pull/102
+[#107]: https://github.com/GIScience/sketch-map-tool/pull/107
+[#116]: https://github.com/GIScience/sketch-map-tool/pull/116